+++ Changelog.md
@@ -8,6 +8,7 @@
 - Show suppressed effects in Effects Panel (since they are shown as token icons).
 - Add "Standard" success result box to Spell, Miracles, Psionics - so that standard benefits can be displayed in test chat cards.
 - Combat Tracker remembers current scroll position when doing "End Turn" on combatants.
+- Add "Apply Active Effect" region behaviour (ready for Foundry V14, and removing the need for the Region Effects module).
 
 ### Translations
 ```